Refactor main application structure with modular utility modules
- Split main.py into multiple focused utility modules
- Created new modules: app_control.py, audio.py, robot_control.py, server.py, ui_helpers.py, updater.py
- Moved related functions and imports to appropriate modules
- Simplified main.py by extracting complex logic into separate utility files
- Improved code organization and maintainability
- Removed redundant global variables and consolidated imports
